Use a common skeleton
